Split src/bin/depclose.rs into a library and a binary.
I needed a program to test driving src/depclose.rs, and then once I had
that it was easy to add other unrelated code into it.  The non-driver
stuff is more related to doing dependency solving so that's all getting
broken out into src/depsolve.rs.  The rest is staying temporarily to
continue testing with, but it can be removed in the future too.
